requirements.txt
attrs==19.1.0
boto3==1.9.217
botocore==1.12.217
certifi==2019.6.16
chalice==1.10.0
chardet==3.0.4
click==6.7
docutils==0.15.2
enum-compat==0.0.2
idna==2.8
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
typing==3.6.4
urllib3==1.25.3
